multi: switch to checked-in generated protobuf/grpc code
Avoids mucking with Git sub modules, and trades away some build
dependencies in favor of a big diff of generated code.
