--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "cusf_sidechain_proto"]
-	path = cusf_sidechain_proto
-	url = https://github.com/LayerTwo-Labs/cusf_sidechain_proto
