+++ .prettierignore
@@ -1,7 +1,4 @@
 # rust build artifacts
 target
 
-# git submodule
-cusf_sidechain_proto
-
 .worktrees
