+++ Cargo.toml
@@ -50,9 +50,7 @@ nonempty = "0.12.0"
 ordermap = "1.2.0"
 parking_lot = "0.12.3"
 prost = "0.14.3"
-prost-build = "0.14.3"
 prost-types = "0.14.3"
-protox = "0.9.0"
 rand = "0.10.1"
 reqwest = { version = "0.13.3", default-features = false, features = ["json"] }
 reserve-port = "2.4.0"
@@ -74,7 +72,6 @@ tokio-util = "0.7.18"
 tonic = "0.14.6"
 tonic-health = "0.14.6"
 tonic-prost = "0.14.5"
-tonic-prost-build = "0.14.6"
 tonic-reflection = "0.14.6"
 # needs to line up with jsonrpsee tower version...
 tower = "0.5.3"
