+++ Justfile
@@ -1,6 +1,11 @@
 default:
     @just --list
 
+# Regenerate checked-in protobuf code under lib/proto/generated/ via buf.
+# The proto source ref is pinned in buf.gen.yaml. Bump it there to upgrade.
+generate:
+    buf generate --clean
+
 clippy:
     cargo clippy --all-targets --all-features --fix --allow-dirty --allow-staged -- --deny warnings
     cargo +nightly clippy -- -A clippy::all -D unqualified_local_imports -Zcrate-attr="feature(unqualified_local_imports)"
