+++ lib/proto/cusf/sidechain.rs
@@ -1,0 +1,5 @@
+#[expect(clippy::allow_attributes, reason = "generated code uses #[allow]")]
+#[path = "../generated/cusf.sidechain.v1.rs"]
+pub mod generated;
+
+pub use self::generated as v1;
