+++ lib/proto/google.rs
@@ -1,0 +1,3 @@
+pub mod protobuf {
+    include!("generated/google.protobuf.rs");
+}
